Выполнил 2 задание
