+++ .gitignore
@@ -13,3 +13,4 @@ node_modules/
 
 # Папка с собранными файлами проекта
 build/
+public/
