+++ src/get-random-integer.js
@@ -1,0 +1,1 @@
+export default (min, max) => Math.floor(Math.random() * (max - min)) + min;
